(wdio-snapshot-service) Update example project
