+++ examples/snapshot/test/specs/complicated_suite.test.ts
@@ -1,7 +1,7 @@
 import { expect, browser, $ } from '@wdio/globals'
 
-describe('My Login application 3', () => {
-    it('should login with valid credentials 3', async () => {
+describe('Level 1 suite', () => {
+    it('test1 1', async () => {
         await browser.url('https://the-internet.herokuapp.com/login')
 
         await $('#username').setValue('tomsmith')
@@ -12,16 +12,17 @@ describe('My Login application 3', () => {
 
         await expect({ a: 'a' }).toMatchSnapshot()
     })
-    describe('My Login application 3 - 2', () => {
-        it('should login with valid credentials 3-2', async () => {
+    describe('Level 2 suite', () => {
+        it('test2 2', async () => {
             await expect({ a: 'a' }).toMatchSnapshot()
         })
-        describe('My Login application 5 - 6', () => {
-            it('should login with valid credentials 5-6', async () => {
+        describe('Level 3 suite', () => {
+            it('test 3', async () => {
                 await expect({ a: 'a' }).toMatchSnapshot()
                 const element = await $('#flash')
                 const html = await element.getHTML()
                 await expect(html).toMatchSnapshot()
+                await expect($('#flash')).toMatchElementSnapshot()
             })
         })
     })
